front: move dependencies
