+++ front/src/applications/editor/components/LinearMetadata/index.tsx
@@ -1,3 +1,2 @@
-export * from './data';
 export * from './FormComponent';
 export * from './FormLineStringLength';
