+++ front/src/common/IntervalsDataViz/data.test.ts
