+++ front/src/common/IntervalsDataViz/dataviz.tsx
