+++ front/src/common/IntervalsDataViz/style.scss
