+++ front/src/common/IntervalsDataViz/utils.ts
